Added dethrace
But I'm not getting audio. And it only uses SDL2. So I'm not sure what's up. Will circle back to it before putting out this way,

+++ dethrace/macsourceports_universal2.sh
@@ -18,44 +18,65 @@ cd ../../${PROJECT_NAME}
 echo git checkout ${GIT_DEFAULT_BRANCH}
 git checkout ${GIT_DEFAULT_BRANCH}
 
-# # fetch the latest 
+# fetch the latest 
 echo git pull
 git pull
 
+# check out the latest release tag
+# echo git checkout tags/${GIT_TAG}
+# git checkout tags/${GIT_TAG}
+
 rm -rf ${BUILT_PRODUCTS_DIR}
 
-# create makefiles with cmake, perform builds with make
-rm -rf ${X86_64_BUILD_FOLDER}
-mkdir ${X86_64_BUILD_FOLDER}
-mkdir -p ${X86_64_BUILD_FOLDER}/${EXECUTABLE_FOLDER_PATH}
-cd ${X86_64_BUILD_FOLDER}
-cmake \
--DCMAKE_OSX_ARCHITECTURES=x86_64 \
--DCMAKE_OSX_DEPLOYMENT_TARGET=10.15 \
--DCMAKE_PREFIX_PATH=/usr/local \
--DCMAKE_INSTALL_PREFIX=/usr/local \
-..
-make -j$NCPU
-cp ${EXECUTABLE_NAME} ${EXECUTABLE_FOLDER_PATH}
+if [ "$1" == "buildserver" ] || [ "$2" == "buildserver" ]; then
+    # mkdir ${BUILT_PRODUCTS_DIR}
+    mkdir -p ${BUILT_PRODUCTS_DIR}/${EXECUTABLE_FOLDER_PATH}
+    cd ${BUILT_PRODUCTS_DIR}
+    cmake "-DCMAKE_OSX_ARCHITECTURES=arm64;x86_64" ..
+    cmake --build . --parallel $NCPU
+    echo cp ${EXECUTABLE_NAME} ${EXECUTABLE_FOLDER_PATH}
+    cp ${EXECUTABLE_NAME} ${EXECUTABLE_FOLDER_PATH}
+    install_name_tool -add_rpath @executable_path/. ${EXECUTABLE_FOLDER_PATH}/${EXECUTABLE_NAME}
+    cp /usr/local/lib/libSDL2-2.0.0.dylib ${EXECUTABLE_FOLDER_PATH}
+    # cp /usr/local/lib/libfmt.11.dylib ${EXECUTABLE_FOLDER_PATH}
+else
+    # create makefiles with cmake, perform builds with make
+    rm -rf ${X86_64_BUILD_FOLDER}
+    mkdir ${X86_64_BUILD_FOLDER}
+    mkdir -p ${X86_64_BUILD_FOLDER}/${EXECUTABLE_FOLDER_PATH}
+    cd ${X86_64_BUILD_FOLDER}
+    cmake \
+    -DCMAKE_OSX_ARCHITECTURES=x86_64 \
+    -DCMAKE_OSX_DEPLOYMENT_TARGET=10.15 \
+    -DCMAKE_PREFIX_PATH=/usr/local \
+    -DCMAKE_INSTALL_PREFIX=/usr/local \
+    ..
+    make -j$NCPU
+    cp ${EXECUTABLE_NAME} ${EXECUTABLE_FOLDER_PATH}
 
-cd ..
-rm -rf ${ARM64_BUILD_FOLDER}
-mkdir ${ARM64_BUILD_FOLDER}
-mkdir -p ${ARM64_BUILD_FOLDER}/${EXECUTABLE_FOLDER_PATH}
-cd ${ARM64_BUILD_FOLDER}
-cmake  \
--DCMAKE_OSX_ARCHITECTURES=arm64 \
--DCMAKE_OSX_DEPLOYMENT_TARGET=10.15 \
--DCMAKE_PREFIX_PATH=/opt/Homebrew \
--DCMAKE_INSTALL_PREFIX=/opt/Homebrew \
-..
-make -j$NCPU
-cp ${EXECUTABLE_NAME} ${EXECUTABLE_FOLDER_PATH}
+    cd ..
+    rm -rf ${ARM64_BUILD_FOLDER}
+    mkdir ${ARM64_BUILD_FOLDER}
+    mkdir -p ${ARM64_BUILD_FOLDER}/${EXECUTABLE_FOLDER_PATH}
+    cd ${ARM64_BUILD_FOLDER}
+    cmake  \
+    -DCMAKE_OSX_ARCHITECTURES=arm64 \
+    -DCMAKE_OSX_DEPLOYMENT_TARGET=10.15 \
+    -DCMAKE_PREFIX_PATH=/opt/Homebrew \
+    -DCMAKE_INSTALL_PREFIX=/opt/Homebrew \
+    ..
+    make -j$NCPU
+    cp ${EXECUTABLE_NAME} ${EXECUTABLE_FOLDER_PATH}
+fi
 
 cd ..
 
 # create the app bundle
-"../MSPBuildSystem/common/build_app_bundle.sh"
+if [ "$1" == "buildserver" ] || [ "$2" == "buildserver" ]; then
+    "../MSPBuildSystem/common/build_app_bundle.sh" "skiplipo"
+else
+    "../MSPBuildSystem/common/build_app_bundle.sh"
+fi
 
 # #sign and notarize
 "../MSPBuildSystem/common/sign_and_notarize.sh" "$1"
